+++ server.js
@@ -2,7 +2,7 @@ const path = require('path');
 const express = require('express');
 const session = require('express-session');
 const exphbs = require('express-handlebars');
-const routes = require('./routes');
+const routes = require('./controllers');
 const sequelize = require('./config/connection');
 const moment = require('moment');
 const helpers = require('./utils/helper');
